extract DonutComponent
