--- src/frontend/ui/home/Bandeau/BandeauDInformation.module.css
@@ -1,3 +1,0 @@
-.bandeau-d-information {
-  margin-top: 2rem;
-}
